16: Login endpoint to query for user by contactid
This commit sets up a /login endpoint that takes an NYCID token,
extracts the email and uses it to look up a Contact in CRM
using the CRM Web API. It returns to the client a ZAP token holding
(signed with) the acquired Contacts contactid.

It also sets up a development scenario where by setting SKIP_AUTH=true
and providing a CRM_IMPOSTER_ID in the environment variables, on startup
the api will automatically trigger the /login endpoint with a mocked NYCID
token, and print the generated ZAP token in the console. This is useful
for bypassing the usual flow of logging in through an email.