Update UI for project lists
This commit updates ui logic to better organize the projects template.  I started to split the project card components based on the ui, but learned from pairing with Andy how to simplify the card styling code, and a strategy for combining these into one component and effectively using ember parameters and conditional logic in the component templates.

Next steps:
- Write the business logic for filtering for “Upcoming” projects, ones that have no pas packege objects
- Implement conditional UI logic for the project cards to show or not show buttons based on the package statuscode
- Update the styling to match the figma wireframes (they’ve diverged a little bit since the github issue 17 was created).

Remaining work on this issue:
- Switch from dummy static data to api request
- Impement breadcrumb component
- Write tests