Add /users endpoint for login functionality
The /users endpoint will take the session's contactId (thanks, auth middleware!)
and look up a CRM Contact for it. It wraps the contact in an array and
passes it to the jsonapi-serializer to return the frontend a JSONAPI compliant
User model. The frontend visits this endpoint after acquiring a ZAP Token.

Mocking Auth and Contact services as providers in app.controller.spec was necessary
for nest tests to run in CI.