Fix landing page sign-in button, redirect to nycid
This commit is a hot fix to have the big orange sign in button on the landing page to redirect users to nyc id.

It copies the logic used in the sign-in button in an index.js controller (left a todo to make this DRY, but wanted to get this hot fix in for demo).