Revert netlify build command syntax for HOST env
Following the Netlify file config for passing branch context environment settings, the redirection when a user logs-in broke.

Looking at network logs, it appears that the HOST environment may be not getting set correctly (it’s Heroku’s develop app for qa and staging branches).

This commit would go back to the syntax we had previously, putting HOST directly in the build command for each context.