83 Handle login errors, like for missing crm contact
This PR modifies the frontend to display error messages
for errors occuring during user login. It adjusts how the
Applicant Portal API sends error responses to support
more robust error reporting on the frontend.

In the frontend, this PR introduces an Ember error page
for the login route. The error page is automatically
displayed by Ember if it detects an exception or
bad response in the route. The exception or bad response
is supplied to the template as the model.

The page also implicitly logs a user out through the DoLogout
component.

In the backend, the PR modifies the AuthService.generateNewToken()
response body to provide an Applicant Portal API-specific error code
 when a Contact cannot be found for the provided email or ID.
This short error code will help the frontend
more robustly display a user-friendly message when no contact is found.