87 Revise text for 'no applicant projects' and 'no active dcp projects whatsoever' scenarios
This commit updates the text shown when there are no applicant (To Do) projects, and also
the text for when there are no Active, Applicant-Only and General Public projects whatsoever
associated with a contact.

the applicant (To Do) projects are a subset of the Active, Applicant-Only and General Public
projects (which are any projects returned from the /projects API endpoint).