Add component actions to add/remove applicant
This commit sets up the team editor component to iterate over a given array of objects and display each fieldset.

It also adds add/remove actions for applicants.  Next step is to define the applicant model & hook up the component to the model so it can dynamically add applicant objects or remove applicant objects.

Outstanding question remains:  where in the data do we have access to the “role” of the applicant?  How do we know which object we get is the “primary applicant”?