110: addresses next to bbl in ProjectGeography component bbl list
When a user searches a bbl or address in the labs-search component,
an object with a bbl and corresponding address is returned. A lookup is
populated with these bbl-address pairs so that addresses can be
displayed next to the bbl in the bbls list.

This commit also provides link to zola on each bbl.