183 Patch applicant geographic state
Since the applicant dcpState field is an option set, we
introduce a dropdown for that field so for users can search and
select for a specific state. In the frontend business logic,
we translate the selected state into its respective option set
code and assign the code to the applicant.dcpState field.