Make generic save button; make more generic pas-form
This refactors parts of the pas-form and most of the save-button to be a little more generic, know a little less about each other.

This simplifies the save-button public API surface, encapsulates some of the button-specific state that pas-form was taking care.

Validations are injectable now, moving the pas-form component to something a little more generic, but not yet.