Make file manager “upsert” new documents
This method is triggered multiple times because it needs to update the reference to documents inside the class.

This change updates-or-creates (“inserts”) a new file manager, depending on if one was already created.

We should get rid of ember-file-upload later so there’s less of a dependency on the service.