Update add and remove applicant state management
Previously, if the applicants array was empty we were adding a blank applicant record so we could render an empty applicant fieldset for the user.   This commit removes that to simplify UX and DX.

When a user clicked “Remove Applicant” button, we were using model.destroyRecord(), which would immediately send a DELETE request to the api server.

In this commit, we use model.deleteRecord() instead, which sets model.isDeleted = true.  This queues the record for deletion when the user clicks “Save” (which ultimately will trigger model.save(), where Ember Data makes the api reqeust to DELETE.