307:  Update add and remove applicant state management (#340)
* Update add and remove applicant state management

Previously, if the applicants array was empty we were adding a blank applicant record so we could render an empty applicant fieldset for the user.   This commit removes that to simplify UX and DX.

When a user clicked “Remove Applicant” button, we were using model.destroyRecord(), which would immediately send a DELETE request to the api server.

In this commit, we use model.deleteRecord() instead, which sets model.isDeleted = true.  This queues the record for deletion when the user clicks “Save” (which ultimately will trigger model.save(), where Ember Data makes the api reqeust to DELETE.

* Update applicant team editor test

* Add displayApplicants getter to fix removed display

This commit uses a displayApplicants getter that filters for only applicants that are not queued for deletion.  This ensures that that user will not see those queued for deletion applicants.

Also updates tests.

* Stop using "new Changeset()"; just call "Changeset()" directly

Based on ember-changeset docs, we should not be calling the changeset
constructor.  Believe this was causing a difficult-to-reproduce browser
bug that would dirty the changeset after the api call to packages finished.

See here:  https://github.com/poteto/ember-changeset#usage