386 load rwcds or pas based on packageType
This PR contains the backend changes to set up the RWCDS form.

The backend has one endpoint to retrieve a package. Previously,
this endpoint relied on a service that assumed all packages
would be sideloaded with a PAS form specifically. This PR
updates the service to sideload either a PAS form or a RWCDS
form based on the package's packageType property.

This is slightly awkward and unsustainable, since each
additional form we build out requires tacking on extra
conditional logic to the package controller and endpoint.
Something to consider is having separate endpoints for
different package types. This would reduce conditional logic
in both the package service and the package controller's JSON
API Serializer.