Introduce StickySidebar component
Componentizes the "Sticky Sidebar" section of PAS Form
for re-use in RWCDS Form and future forms. Assumes navigation
items are a fixture of the component, while consumers of
StickySidebar can use the block form to yield things inside
the sidebar, like Save or Submit buttons which are wired to
the Saveable Form.