410 Wire up RWCDS Save button
This commit sets up the form Save button, along with the necessary
save function within the Edit component to trigger a package save.
That follows the same pattern as in PAS Form.

More critically, it modifies the Package model's save function to
conditionally save either the package's child PAS Form or RWCDS
form, based on dcpPackagetype. This is crucial because we work on
the assumption that a Package only represents either a PAS form or
a RWCDS form.

This commit also needed to resolve the conflict between the backend's
Project endpoint delivering sideloaded Packages with the dcPackagetype
property already translated, and the Package endpoint not yet
translating dcpPackagetype. This was causing confusion in the frontend
app and tests as to whether the property should be regarded as a
string or a number.