222 Prefill RWCDS description fields from PAS
Modifies the Package Service getPackage() function
to check for if two description fields are
blank on retrieved RWCDS forms. If so, it queries
for the PAS Form on the latest PAS Package revision
(on the same Project as the RWCDS), and prefills
those the RWCDS form fields from the PAS Form fields.
The pertinent fields are specified in Issue 222.