526 Flatten nested response objects and mold to JSON API format
Rework the HTTP Exception Filter to mold Exceptions to JSON API
compliant error objects before returning them to the client.

Also, response objects from HttpExceptions may be recursively
nested to reflect a stack trace (each level is a layer in the stack).
This flattens them out so that each response
is a top-level item in the errors array.