Auth service looks up/creates record (upsert)
Adds behavior in auth to react to user’s status within the nyc.id system.

If the contact is found, by GUID, prefer that. If not, then check if there’s a record from the e-mail. If neither case, create a new record with the e-mail/guid.

Lastly, in any case, make sure that the contact record’s GUID field is updated.