12354: fix bug where deleted records were not being saved
A while back, we switched from destroyRecord to deleteRecord. Since
then, when we looped through the records with dirty attributes and only
saved those records, we were not taking into account records that had
been deleted -- these records did not show up in the array and were
therefore not saved. This commit passes the records that are currently
being "removed" to the package model to be saved.