12354: refactor deleting bbls
A while back, we switched from destroyRecord to deleteRecord.
Since then, when we looped through the records with dirty attributes
and only saved those records, we were not taking into account records
that had been deleted -- these records did not show up in the array
and were therefore not saved. This PR passes the records that are
currently being "removed" to the package model to be saved.

This commit refactors the way that we are removing bbls to match
applicants and relatedActions. It also handles resetting the
recordsToBeDeleted object so that the save button is correctly
enabled and disabled.