12395 Draft LU Zoning Map Amendment
This PR sets an example for calling *-fieldset components going forward. Notice how the zoning-map-amendment-fieldset component is called without the @zoningMapChange argument, since that argument is redundant and confusing.
Similarly this PR sets an example for calling the @ remove<fieldset> actions. Previously in other sections we passed an extra, unnecessary argument to the @ remove<fieldset> action inside the *-fieldset file (at the remove button).
We should later assess whether the ZONING_MAP_CHANGE_OPTIONSET.DCPEXISTINGZONINGDISTRICTVALUE
optionset is frequently growing/shrinking. If so, we should pull this optionset down on app load. I created a ticket here: AB#12865
Another area we are pushing up against the "saveable form" api: the Q.Label component doesn't create working data-test hooks for power selects. Power select need a redundant <label> to house working test selectors.
In CRM, the singular entity name for a Zoning Map Change is dcp_zoningmapchanges, and plural is dcp_zoningmapchangeses. But we hide most of the app from this detail. It only affects the direct, low-level calls to the CRM Web API.